requirements.txt
streamlit>=1.28.0
groq>=0.4.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit-mic-recorder>=0.0.8